+++ contracts/protocols/lending/aave/v3.6/deployments/utilities/DeployUtils.sol
@@ -1,73 +1,67 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.0;
 
-import 'forge-std/StdJson.sol';
-import 'forge-std/Vm.sol';
-import '../interfaces/IMarketReportTypes.sol';
-import '@crane/contracts/protocols/lending/aave/v3.6/dependencies/solidity-utils/contracts/transparent-proxy/TransparentProxyFactory.sol';
+import "forge-std/StdJson.sol";
+import "forge-std/Vm.sol";
+import "../interfaces/IMarketReportTypes.sol";
+import "@crane/contracts/protocols/lending/aave/v3.6/dependencies/solidity-utils/contracts/transparent-proxy/TransparentProxyFactory.sol";
 
-import {BetterEfficientHashLib} from '@crane/contracts/utils/BetterEfficientHashLib.sol';
+import {BetterEfficientHashLib} from "@crane/contracts/utils/BetterEfficientHashLib.sol";
 
 contract DeployUtils {
-  
     using BetterEfficientHashLib for bytes;
 
     using stdJson for string;
 
-    Vm private constant vm = Vm(address(bytes20(uint160(uint256(keccak256('hevm cheat code'))))));
+    Vm private constant vm = Vm(address(bytes20(uint160(uint256(keccak256("hevm cheat code"))))));
 
     function _deployFromArtifacts(string memory contractPath) internal returns (address deployment) {
         bytes memory bytecode = abi.encodePacked(vm.getCode(contractPath));
         assembly {
-          deployment := create(0, add(bytecode, 0x20), mload(bytecode))
+            deployment := create(0, add(bytecode, 0x20), mload(bytecode))
         }
 
         return deployment;
     }
 
-    function _deployFromArtifactsWithBroadcast(
-        string memory contractPath
-    ) internal returns (address deployment) {
+    function _deployFromArtifactsWithBroadcast(string memory contractPath) internal returns (address deployment) {
         bytes memory bytecode = abi.encodePacked(vm.getCode(contractPath));
         vm.broadcast();
         assembly {
-          deployment := create(0, add(bytecode, 0x20), mload(bytecode))
+            deployment := create(0, add(bytecode, 0x20), mload(bytecode))
         }
 
         return deployment;
     }
 
-    function _deployFromArtifactsWithBroadcast(
-        string memory contractPath,
-        bytes memory args
-    ) internal returns (address deployment) {
+    function _deployFromArtifactsWithBroadcast(string memory contractPath, bytes memory args)
+        internal
+        returns (address deployment)
+    {
         bytes memory bytecode = abi.encodePacked(vm.getCode(contractPath), args);
 
         vm.broadcast();
         assembly {
-          deployment := create(0, add(bytecode, 0x20), mload(bytecode))
+            deployment := create(0, add(bytecode, 0x20), mload(bytecode))
         }
 
         return deployment;
     }
 
-    function _deployFromArtifacts(
-        string memory contractPath,
-        bytes memory args
-    ) internal returns (address deployment) {
+    function _deployFromArtifacts(string memory contractPath, bytes memory args) internal returns (address deployment) {
         bytes memory bytecode = abi.encodePacked(vm.getCode(contractPath), args);
         assembly {
-          deployment := create(0, add(bytecode, 0x20), mload(bytecode))
+            deployment := create(0, add(bytecode, 0x20), mload(bytecode))
         }
 
         return deployment;
     }
 
-    function getCreate2Address(
-        string memory contractPath,
-        bytes memory args,
-        bytes32 _salt
-    ) public view returns (address) {
+    function getCreate2Address(string memory contractPath, bytes memory args, bytes32 _salt)
+        public
+        view
+        returns (address)
+    {
         bytes memory bytecode = abi.encodePacked(vm.getCode(contractPath), args);
 
         // bytes32 hash = keccak256(
@@ -75,6 +69,6 @@ contract DeployUtils {
         // );
         bytes32 hash = abi.encodePacked(bytes1(0xff), address(this), _salt, keccak256(bytecode))._hash();
 
-        return address(uint160(uint(hash)));
+        return address(uint160(uint256(hash)));
     }
 }
