+++ contracts/protocols/perps/pendle/interfaces/Thena/IThenaPair.sol
@@ -3,26 +3,30 @@
 pragma solidity ^0.8.0;
 
 interface IThenaPair {
-    function metadata() external view returns (uint dec0, uint dec1, uint r0, uint r1, bool st, address t0, address t1);
-    function claimFees() external returns (uint, uint);
+    function metadata()
+        external
+        view
+        returns (uint256 dec0, uint256 dec1, uint256 r0, uint256 r1, bool st, address t0, address t1);
+    function claimFees() external returns (uint256, uint256);
     function tokens() external view returns (address, address);
     function token0() external view returns (address);
     function token1() external view returns (address);
-    function transferFrom(address src, address dst, uint amount) external returns (bool);
-    function permit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
-    function burn(address to) external returns (uint amount0, uint amount1);
-    function mint(address to) external returns (uint liquidity);
-    function getReserves() external view returns (uint _reserve0, uint _reserve1, uint _blockTimestampLast);
-    function getAmountOut(uint, address) external view returns (uint);
+    function transferFrom(address src, address dst, uint256 amount) external returns (bool);
+    function permit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s)
+        external;
+    function swap(uint256 amount0Out, uint256 amount1Out, address to, bytes calldata data) external;
+    function burn(address to) external returns (uint256 amount0, uint256 amount1);
+    function mint(address to) external returns (uint256 liquidity);
+    function getReserves() external view returns (uint256 _reserve0, uint256 _reserve1, uint256 _blockTimestampLast);
+    function getAmountOut(uint256, address) external view returns (uint256);
 
     function name() external view returns (string memory);
     function symbol() external view returns (string memory);
-    function totalSupply() external view returns (uint);
+    function totalSupply() external view returns (uint256);
     function decimals() external view returns (uint8);
 
-    function claimable0(address _user) external view returns (uint);
-    function claimable1(address _user) external view returns (uint);
+    function claimable0(address _user) external view returns (uint256);
+    function claimable1(address _user) external view returns (uint256);
 
     function isStable() external view returns (bool);
 }
