+++ contracts/protocols/tokens/stable/frax/ERC20/IWETH.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.35;
 
 interface IWETH {
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function transferFrom(address src, address dst, uint wad) external returns (bool);
-    function withdraw(uint) external;
+    function transfer(address to, uint256 value) external returns (bool);
+    function transferFrom(address src, address dst, uint256 wad) external returns (bool);
+    function withdraw(uint256) external;
 }
